requirements.txt
fastapi
pydantic
openai
uvicorn
python-dotenv
pymongo
python-multipart
